+++ LightPattern.h
@@ -17,7 +17,7 @@ class LightPattern: public Event
     
   public:
     LightPattern(Clock* pClock, LedController* pLedController);
-    LightPattern(int duration, Clock* pClock, LedController* pLedController);
+    LightPattern(unsigned long duration, Clock* pClock, LedController* pLedController);
     LightPattern(LightPatternOptions lightPatternOptions, Clock* pClock, LedController* pLedController);
     void display();
     void display(int frequency);
